server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-groq

# Vectorstore (Cloud)
pinecone-client

# Embeddings
google-generativeai
sentence-transformers

# Local Vector DB (optional fallback)
faiss-cpu

# PDF + OCR Processing
pymupdf            # (fitz) used for PDF + image extraction
pypdf              # lightweight fallback parser
pillow             # image handling

# Tesseract OCR (main OCR engine)
pytesseract
opencv-python       # image preprocessing for OCR

# Optional OCR for handwriting / blurry images
easyocr

# Unstructured (still useful for HTML, email, docx, etc.)
unstructured
unstructured-inference
unstructured[pdf]

# File type detection (Windows safe)
python-magic-bin

# Environment Variables
python-dotenv

# Utilities
pydantic
requests
tqdm

# Database + Auth (if needed in your backend)
pymongo
bcrypt


#  uv pip install -r requirements.txt